+++ src/init.lua
@@ -1,3 +1,4 @@
+-- Compiled with L+ C Edition
 -- Janitor
 -- Original by Validark
 -- Modifications by pobammer
@@ -6,6 +7,7 @@
 -- Cleanup edge cases fixed by codesenseAye.
 
 local GetPromiseLibrary = require(script.GetPromiseLibrary)
+local RbxScriptConnection = require(script.RbxScriptConnection)
 local Symbol = require(script.Symbol)
 local FoundPromiseLibrary, Promise = GetPromiseLibrary()
 
@@ -14,6 +16,8 @@ local LinkToInstanceIndex = Symbol("LinkToInstanceIndex")
 
 local METHOD_NOT_FOUND_ERROR = "Object %s doesn't have method %s, are you sure you want to add it? Traceback: %s"
 local NOT_A_PROMISE = "Invalid argument #1 to 'Janitor:AddPromise' (Promise expected, got %s (%s))"
+
+type RbxScriptConnection = RbxScriptConnection.RbxScriptConnection
 
 --[=[
 	Janitor is a light-weight, flexible object for cleaning up connections, instances, or anything. This implementation covers all use cases,
@@ -28,10 +32,29 @@ Janitor.CurrentlyCleaning = true
 Janitor[IndicesReference] = nil
 Janitor.__index = Janitor
 
+--[=[
+	@prop CurrentlyCleaning boolean
+	@within Janitor
+
+	Whether or not the Janitor is currently cleaning up. Will be `nil` if it is.
+]=]
+
 local TypeDefaults = {
 	["function"] = true;
+	thread = true;
 	RBXScriptConnection = "Disconnect";
 }
+
+--[=[
+	Instantiates a new Janitor object.
+	@return Janitor
+]=]
+function Janitor.new()
+	return setmetatable({
+		CurrentlyCleaning = false;
+		[IndicesReference] = nil;
+	}, Janitor)
+end
 
 --[=[
 	Determines if the passed object is a Janitor. This checks the metatable directly.
@@ -47,14 +70,16 @@ type StringOrTrue = string | boolean
 
 --[=[
 	Adds an `Object` to Janitor for later cleanup, where `MethodName` is the key of the method within `Object` which should be called at cleanup time.
-	If the `MethodName` is `true` the `Object` itself will be called instead. If passed an index it will occupy a namespace which can be `Remove()`d or overwritten.
-	Returns the `Object`.
+	If the `MethodName` is `true` the `Object` itself will be called if it's a function or have `task.cancel` called on it if it is a thread. If passed
+	an index it will occupy a namespace which can be `Remove()`d or overwritten. Returns the `Object`.
 
 	:::info
 	Objects not given an explicit `MethodName` will be passed into the `typeof` function for a very naive typecheck.
-	RBXConnections will be assigned to "Disconnect", functions will be assigned to `true`, and everything else will default to "Destroy".
+	RBXConnections will be assigned to "Disconnect", functions and threads will be assigned to `true`, and everything else will default to "Destroy".
 	Not recommended, but hey, you do you.
 	:::
+
+	### Luau:
 
 	```lua
 	local Workspace = game:GetService("Workspace")
@@ -69,6 +94,14 @@ type StringOrTrue = string | boolean
 	-- Queue function to be called with `true` MethodName
 	Obliterator:Add(print, true)
 
+	-- Close a thread.
+	Obliterator:Add(task.defer(function()
+		while true do
+			print("Running!")
+			task.wait(0.5)
+		end
+	end), true)
+
 	-- This implementation allows you to specify behavior for any object
 	Obliterator:Add(TweenService:Create(Part, TweenInfo.new(1), {Size = Vector3.new(1, 1, 1)}), "Cancel")
 
@@ -76,6 +109,8 @@ type StringOrTrue = string | boolean
 	-- If "CurrentTween" already exists, it will call :Remove("CurrentTween") before writing
 	Obliterator:Add(TweenService:Create(Part, TweenInfo.new(1), {Size = Vector3.new(1, 1, 1)}), "Destroy", "CurrentTween")
 	```
+
+	### TypeScript:
 
 	```ts
 	import { Workspace, TweenService } from "@rbxts/services";
@@ -90,6 +125,14 @@ type StringOrTrue = string | boolean
 	// Queue function to be called with `true` MethodName
 	Obliterator.Add(print, true);
 
+	// Close a thread.
+	Obliterator.Add(task.defer(() => {
+		while (true) {
+			print("Running!");
+			task.wait(0.5);
+		}
+	}), true);
+
 	// This implementation allows you to specify behavior for any object
 	Obliterator.Add(TweenService.Create(Part, new TweenInfo(1), {Size: new Vector3(1, 1, 1)}), "Cancel");
 
@@ -103,7 +146,7 @@ type StringOrTrue = string | boolean
 	@param Index? any -- The index that can be used to clean up the object manually.
 	@return T -- The object that was passed as the first argument.
 ]=]
-function Janitor:Add(Object: any, MethodName: StringOrTrue?, Index: any?): any
+function Janitor:Add<T>(Object: T, MethodName: StringOrTrue?, Index: any?): T
 	if Index then
 		self:Remove(Index)
 
@@ -116,17 +159,20 @@ function Janitor:Add(Object: any, MethodName: StringOrTrue?, Index: any?): any
 		This[Index] = Object
 	end
 
-	MethodName = MethodName or TypeDefaults[typeof(Object)] or "Destroy"
-	if type(Object) ~= "function" and not Object[MethodName] then
-		warn(string.format(METHOD_NOT_FOUND_ERROR, tostring(Object), tostring(MethodName), debug.traceback(nil :: any, 2)))
+	local TypeOf = typeof(Object)
+	local NewMethodName = MethodName or TypeDefaults[TypeOf] or "Destroy"
+	if TypeOf ~= "function" and TypeOf ~= "thread" and not (Object :: any)[NewMethodName] then
+		warn(string.format(METHOD_NOT_FOUND_ERROR, tostring(Object), tostring(NewMethodName), debug.traceback(nil :: any, 2)))
 	end
 
-	self[Object] = MethodName
+	self[Object] = NewMethodName
 	return Object
 end
 
 --[=[
 	Adds a [Promise](https://github.com/evaera/roblox-lua-promise) to the Janitor. If the Janitor is cleaned up and the Promise is not completed, the Promise will be cancelled.
+
+	### Luau:
 
 	```lua
 	local Obliterator = Janitor.new()
@@ -134,6 +180,8 @@ end
 	task.wait(1)
 	Obliterator:Cleanup()
 	```
+
+	### TypeScript:
 
 	```ts
 	import { Janitor } from "@rbxts/janitor";
@@ -178,11 +226,15 @@ end
 --[=[
 	Cleans up whatever `Object` was set to this namespace by the 3rd parameter of [Janitor.Add](#Add).
 
+	### Luau:
+
 	```lua
 	local Obliterator = Janitor.new()
 	Obliterator:Add(workspace.Baseplate, "Destroy", "Baseplate")
 	Obliterator:Remove("Baseplate")
 	```
+
+	### TypeScript:
 
 	```ts
 	import { Workspace } from "@rbxts/services";
@@ -207,7 +259,11 @@ function Janitor:Remove(Index: any)
 
 			if MethodName then
 				if MethodName == true then
-					Object()
+					if type(Object) == "function" then
+						Object()
+					else
+						task.cancel(Object)
+					end
 				else
 					local ObjectMethod = Object[MethodName]
 					if ObjectMethod then
@@ -226,13 +282,96 @@ function Janitor:Remove(Index: any)
 end
 
 --[=[
+	Cleans up multiple objects at once.
+
+	### Luau:
+
+	```lua
+	local Obliterator = Janitor.new()
+	Obliterator:Add(function()
+		print("Removed One")
+	end, true, "One")
+
+	Obliterator:Add(function()
+		print("Removed Two")
+	end, true, "Two")
+
+	Obliterator:Add(function()
+		print("Removed Three")
+	end, true, "Three")
+
+	Obliterator:RemoveList("One", "Two", "Three") -- Prints "Removed One", "Removed Two", and "Removed Three"
+	```
+
+	### TypeScript:
+
+	```ts
+	import { Janitor } from "@rbxts/janitor";
+
+	type NoOp = () => void
+
+	const Obliterator = new Janitor<{ One: NoOp, Two: NoOp, Three: NoOp }>();
+	Obliterator.Add(() => print("Removed One"), true, "One");
+	Obliterator.Add(() => print("Removed Two"), true, "Two");
+	Obliterator.Add(() => print("Removed Three"), true, "Three");
+
+	Obliterator.RemoveList("One", "Two", "Three"); // Prints "Removed One", "Removed Two", and "Removed Three"
+	```
+
+	@param ... any -- The indices you want to remove.
+	@return Janitor
+]=]
+function Janitor:RemoveList(...: any)
+	local This = self[IndicesReference]
+	if This then
+		local Length = select("#", ...)
+		if Length == 1 then
+			return self:Remove(...)
+		else
+			for Index = 1, Length do
+				-- MACRO
+				local Object = This[select(Index, ...)]
+				if Object then
+					local MethodName = self[Object]
+
+					if MethodName then
+						if MethodName == true then
+							if type(Object) == "function" then
+								Object()
+							else
+								task.cancel(Object)
+							end
+						else
+							local ObjectMethod = Object[MethodName]
+							if ObjectMethod then
+								ObjectMethod(Object)
+							end
+						end
+
+						self[Object] = nil
+					end
+
+					This[Index] = nil
+				end
+			end
+		end
+	end
+
+	return self
+end
+
+--[=[
 	Gets whatever object is stored with the given index, if it exists. This was added since Maid allows getting the task using `__index`.
+
+	### Luau:
 
 	```lua
 	local Obliterator = Janitor.new()
 	Obliterator:Add(workspace.Baseplate, "Destroy", "Baseplate")
 	print(Obliterator:Get("Baseplate")) -- Returns Baseplate.
 	```
+
+	### TypeScript:
 
 	```ts
 	import { Workspace } from "@rbxts/services";
@@ -248,16 +387,12 @@ end
 ]=]
 function Janitor:Get(Index: any): any?
 	local This = self[IndicesReference]
-	if This then
-		return This[Index]
-	else
-		return nil
-	end
+	return if This then This[Index] else nil
 end
 
 local function GetFenv(self)
 	return function()
-		for Object, MethodName in pairs(self) do
+		for Object, MethodName in next, self do
 			if Object ~= IndicesReference then
 				return Object, MethodName
 			end
@@ -269,10 +404,14 @@ end
 	Calls each Object's `MethodName` (or calls the Object if `MethodName == true`) and removes them from the Janitor. Also clears the namespace.
 	This function is also called when you call a Janitor Object (so it can be used as a destructor callback).
 
+	### Luau:
+
 	```lua
 	Obliterator:Cleanup() -- Valid.
 	Obliterator() -- Also valid.
 	```
+
+	### TypeScript:
 
 	```ts
 	Obliterator.Cleanup()
@@ -287,7 +426,11 @@ function Janitor:Cleanup()
 
 		while Object and MethodName do -- changed to a while loop so that if you add to the janitor inside of a callback it doesn't get untracked (instead it will loop continuously which is a lot better than a hard to pindown edgecase)
 			if MethodName == true then
-				Object()
+				if type(Object) == "function" then
+					Object()
+				else
+					task.cancel(Object)
+				end
 			else
 				local ObjectMethod = Object[MethodName]
 				if ObjectMethod then
@@ -313,7 +456,7 @@ end
 	Calls [Janitor.Cleanup](#Cleanup) and renders the Janitor unusable.
 
 	:::warning
-	Running this will make any attempts to call a function of Janitor error.
+	Running this will make any further attempts to call a method of Janitor error.
 	:::
 ]=]
 function Janitor:Destroy()
@@ -325,42 +468,13 @@ end
 Janitor.__call = Janitor.Cleanup
 
 --[=[
-	A wrapper for an `RBXScriptConnection`. Makes the Janitor clean up when the instance is destroyed. This was created by Corecii.
-
-	@class RbxScriptConnection
-	@__index RbxScriptConnection
-]=]
-local RbxScriptConnection = {}
-RbxScriptConnection.Connected = true
-RbxScriptConnection.__index = RbxScriptConnection
-
---[=[
-	Disconnects the Signal.
-]=]
-function RbxScriptConnection:Disconnect()
-	if self.Connected then
-		self.Connected = false
-		self.Connection:Disconnect()
-	end
-end
-
-function RbxScriptConnection._new(RBXScriptConnection: RBXScriptConnection)
-	return setmetatable({
-		Connection = RBXScriptConnection;
-	}, RbxScriptConnection)
-end
-
-function RbxScriptConnection:__tostring()
-	return "RbxScriptConnection<" .. tostring(self.Connected) .. ">"
-end
-
-type RbxScriptConnection = typeof(RbxScriptConnection._new(game:GetPropertyChangedSignal("ClassName"):Connect(function() end)))
-
---[=[
 	"Links" this Janitor to an Instance, such that the Janitor will `Cleanup` when the Instance is `Destroyed()` and garbage collected.
 	A Janitor may only be linked to one instance at a time, unless `AllowMultiple` is true. When called with a truthy `AllowMultiple` parameter,
 	the Janitor will "link" the Instance without overwriting any previous links, and will also not be overwritable.
 	When called with a falsy `AllowMultiple` parameter, the Janitor will overwrite the previous link which was also called with a falsy `AllowMultiple` parameter, if applicable.
+	This returns a mock `RBXScriptConnection` (see: [RbxScriptConnection](/api/RbxScriptConnection)).
+
+	### Luau:
 
 	```lua
 	local Obliterator = Janitor.new()
@@ -376,6 +490,8 @@ type RbxScriptConnection = typeof(RbxScriptConnection._new(game:GetPropertyChang
 	end
 	```
 
+	### TypeScript:
+
 	```ts
 	import { Janitor } from "@rbxts/janitor";
 
@@ -388,8 +504,6 @@ type RbxScriptConnection = typeof(RbxScriptConnection._new(game:GetPropertyChang
 		Folder.Destroy();
 	}
 	```
-
-	This returns a mock `RBXScriptConnection` (see: [RbxScriptConnection](#RbxScriptConnection)).
 
 	@param Object Instance -- The instance you want to link the Janitor to.
 	@param AllowMultiple? boolean -- Whether or not to allow multiple links on the same Janitor.
@@ -452,16 +566,10 @@ function Janitor:LinkToInstances(...: Instance)
 	return ManualCleanup
 end
 
---[=[
-	Instantiates a new Janitor object.
-	@return Janitor
-]=]
-function Janitor.new()
-	return setmetatable({
-		CurrentlyCleaning = false;
-		[IndicesReference] = nil;
-	}, Janitor)
+function Janitor:__tostring()
+	return "Janitor"
 end
 
 export type Janitor = typeof(Janitor.new())
+table.freeze(Janitor)
 return Janitor